Change type of temporary vars to properly clip. Fixes SMT problem.

+++ apu/bapu/smp/smp.hpp
@@ -50,7 +50,7 @@ public:
   unsigned opcode_number;
   unsigned opcode_cycle;
 
-  unsigned rd, wr, dp, sp, ya, bit;
+  uint16 rd, wr, dp, sp, ya, bit;
 
   struct Regs {
     uint16 pc;
